+++ controllers/state_of_the_world.go
@@ -8,16 +8,15 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/google/go-cmp/cmp"
 	authorinov1beta1 "github.com/kuadrant/authorino-operator/api/v1beta1"
+	"github.com/kuadrant/limitador-operator/api/v1alpha1"
 	"github.com/kuadrant/policy-machinery/controller"
 	"github.com/kuadrant/policy-machinery/machinery"
-	"github.com/samber/lo"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/utils/env"
-	"k8s.io/utils/ptr"
 	ctrlruntime "sigs.k8s.io/controller-runtime"
 	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 
@@ -25,6 +24,7 @@ import (
 	kuadrantv1beta1 "github.com/kuadrant/kuadrant-operator/api/v1beta1"
 	kuadrantv1beta2 "github.com/kuadrant/kuadrant-operator/api/v1beta2"
 	"github.com/kuadrant/kuadrant-operator/pkg/library/kuadrant"
+	"github.com/kuadrant/kuadrant-operator/pkg/resources"
 )
 
 var (
@@ -49,6 +49,7 @@ func NewPolicyMachineryController(manager ctrlruntime.Manager, client *dynamic.D
 		controller.WithRunnable("ratelimitpolicy watcher", controller.Watch(&kuadrantv1beta2.RateLimitPolicy{}, kuadrantv1beta2.RateLimitPoliciesResource, metav1.NamespaceAll)),
 		controller.WithRunnable("topology configmap watcher", controller.Watch(&corev1.ConfigMap{}, controller.ConfigMapsResource, operatorNamespace, controller.FilterResourcesByLabel[*corev1.ConfigMap](fmt.Sprintf("%s=true", kuadrant.TopologyLabel)))),
 		controller.WithRunnable("authorino watcher", controller.Watch(&authorinov1beta1.Authorino{}, kuadrantv1beta1.AuthorinoResource, metav1.NamespaceAll)),
+		controller.WithRunnable("limitador watcher", controller.Watch(&v1alpha1.Limitador{}, kuadrantv1beta1.LimitadorResource, metav1.NamespaceAll)),
 		controller.WithPolicyKinds(
 			kuadrantv1alpha1.DNSPolicyKind,
 			kuadrantv1alpha1.TLSPolicyKind,
@@ -59,10 +60,12 @@ func NewPolicyMachineryController(manager ctrlruntime.Manager, client *dynamic.D
 			kuadrantv1beta1.KuadrantKind,
 			ConfigMapGroupKind,
 			kuadrantv1beta1.AuthorinoKind,
+			kuadrantv1beta1.LimitadorKind,
 		),
 		controller.WithObjectLinks(
 			kuadrantv1beta1.LinkKuadrantToGatewayClasses,
 			kuadrantv1beta1.LinkKuadrantToAuthorino,
+			kuadrantv1beta1.LinkKuadrantToLimitador,
 		),
 		controller.WithReconcile(buildReconciler(client)),
 	}
@@ -79,115 +82,11 @@ func buildReconciler(client *dynamic.DynamicClient) controller.ReconcileFunc {
 			},
 		}).Run,
 		Tasks: []controller.ReconcileFunc{
-			NewAuthorinoCrReconciler(client).Subscription().Reconcile,
+			resources.NewAuthorinoCrReconciler(client).Subscription().Reconcile,
+			resources.NewLimitadorCrReconciler(client).Subscription().Reconcile,
 		},
 	}
 	return reconciler.Run
-}
-
-type AuthorinoCrReconciler struct {
-	Client *dynamic.DynamicClient
-}
-
-func NewAuthorinoCrReconciler(client *dynamic.DynamicClient) *AuthorinoCrReconciler {
-	return &AuthorinoCrReconciler{Client: client}
-}
-
-func (r *AuthorinoCrReconciler) Subscription() *controller.Subscription {
-	return &controller.Subscription{
-		ReconcileFunc: r.Reconcile,
-		Events: []controller.ResourceEventMatcher{
-			{Kind: ptr.To(kuadrantv1beta1.KuadrantKind), EventType: ptr.To(controller.CreateEvent)},
-			{Kind: ptr.To(kuadrantv1beta1.AuthorinoKind), EventType: ptr.To(controller.DeleteEvent)},
-		},
-	}
-}
-
-func (r *AuthorinoCrReconciler) Reconcile(ctx context.Context, _ []controller.ResourceEvent, topology *machinery.Topology, _ error) {
-	logger := controller.LoggerFromContext(ctx).WithName("AuthorinoCrReconciler")
-	logger.Info("reconciling authorino resource", "status", "started")
-	defer logger.Info("reconciling authorino resource", "status", "completed")
-
-	kobjs := lo.FilterMap(topology.Objects().Roots(), func(item machinery.Object, _ int) (*kuadrantv1beta1.Kuadrant, bool) {
-		if item.GroupVersionKind().Kind == kuadrantv1beta1.KuadrantKind.Kind {
-			return item.(*kuadrantv1beta1.Kuadrant), true
-		}
-		return nil, false
-	})
-	if len(kobjs) == 0 {
-		logger.Info("no kuadrant resources found", "status", "skipping")
-		return
-	}
-	if len(kobjs) > 1 {
-		logger.Error(fmt.Errorf("multiple Kuadrant resources found"), "cannot select root Kuadrant resource", "status", "error")
-	}
-	kobj := kobjs[0]
-
-	if kobj.GetDeletionTimestamp() != nil {
-		logger.Info("root kuadrant marked for deletion", "status", "skipping")
-		return
-	}
-
-	aobjs := lo.FilterMap(topology.Objects().Objects().Items(), func(item machinery.Object, _ int) (machinery.Object, bool) {
-		if item.GroupVersionKind().Kind == kuadrantv1beta1.AuthorinoKind.Kind {
-			return item, true
-		}
-		return nil, false
-	})
-
-	if len(aobjs) > 0 {
-		logger.Info("authorino resource already exists, no need to create", "status", "skipping")
-		return
-	}
-
-	authorino := &authorinov1beta1.Authorino{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Authorino",
-			APIVersion: "operator.authorino.kuadrant.io/v1beta1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "authorino",
-			Namespace: kobj.Namespace,
-			OwnerReferences: []metav1.OwnerReference{
-				{
-					APIVersion:         kobj.GroupVersionKind().GroupVersion().String(),
-					Kind:               kobj.GroupVersionKind().Kind,
-					Name:               kobj.Name,
-					UID:                kobj.UID,
-					BlockOwnerDeletion: ptr.To(true),
-					Controller:         ptr.To(true),
-				},
-			},
-		},
-		Spec: authorinov1beta1.AuthorinoSpec{
-			ClusterWide:            true,
-			SupersedingHostSubsets: true,
-			Listener: authorinov1beta1.Listener{
-				Tls: authorinov1beta1.Tls{
-					Enabled: ptr.To(false),
-				},
-			},
-			OIDCServer: authorinov1beta1.OIDCServer{
-				Tls: authorinov1beta1.Tls{
-					Enabled: ptr.To(false),
-				},
-			},
-		},
-	}
-
-	unstructuredAuthorino, err := controller.Destruct(authorino)
-	if err != nil {
-		logger.Error(err, "failed to destruct authorino", "status", "error")
-	}
-	logger.Info("creating authorino resource", "status", "processing")
-	_, err = r.Client.Resource(kuadrantv1beta1.AuthorinoResource).Namespace(authorino.Namespace).Create(ctx, unstructuredAuthorino, metav1.CreateOptions{})
-	if err != nil {
-		if errors.IsAlreadyExists(err) {
-			logger.Info("already created authorino resource", "status", "acceptable")
-		} else {
-			logger.Error(err, "failed to create authorino resource", "status", "error")
-		}
-	}
 }
 
 type TopologyFileReconciler struct {
